Changing the location of the tanh function in the numpy frontend from the trigonometric_functions.py file to the hyperbolic_functions.py as well as its to test_hyperbolic_functions.py
