+++ ivy/functional/frontends/numpy/mathematical_functions/trigonometric_functions.py
@@ -121,29 +121,6 @@ arcsin.unsupported_dtypes = {"torch": ("float16",)}
 
 
 @from_zero_dim_arrays_to_float
-def tanh(
-    x,
-    /,
-    out=None,
-    *,
-    where=True,
-    casting="same_kind",
-    order="k",
-    dtype=None,
-    subok=True,
-):
-    if dtype:
-        x = ivy.astype(ivy.array(x), ivy.as_ivy_dtype(dtype))
-    ret = ivy.tanh(x, out=out)
-    if ivy.is_array(where):
-        ret = ivy.where(where, ret, ivy.default(out, ivy.zeros_like(ret)), out=out)
-    return ret
-
-
-tanh.unsupported_dtypes = {"torch": ("float16",)}
-
-
-@from_zero_dim_arrays_to_float
 def arctan(
     x,
     /,
